[#2] Add unit tests for CPU, Disk, and GPU metrics clients; enhance metrics handling
- Implemented unit tests for CpuTemperatureMetricsClient, DiskHealthMetricsClient, DiskSpeedMetricsClient, and GpuMetricsClient to ensure proper functionality and error handling.
- Added a comprehensive documentation file detailing the status of various metrics fields across platforms.
- Enhanced the CpuTemperatureMetricsClient to handle WMI queries for Windows and improved error messaging.
- Updated DiskHealthMetricsClient and DiskSpeedMetricsClient to utilize dependency injection for providers.
- Modified GpuMetrics to use double for TotalMemoryMB for better precision.
- Improved macOS GPU metrics retrieval using system_profiler and added JSON parsing for display data.

+++ src/RapidStreamer.BuildingBlocks.Infrastructure/SystemResourceMonitor/Metrics/Gpu/GpuMetrics.cs
@@ -33,7 +33,7 @@ public record GpuMetrics : IMetrics
     /// <summary>
     /// Total GPU memory in MB. Null if not supported.
     /// </summary>
-    public long? TotalMemoryMB { get; init; }
+    public double? TotalMemoryMB { get; init; }
 
     /// <summary>
     /// Used GPU memory in MB. Null if not supported.
